% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atRegions.R
\name{atRegions}
\alias{atRegions}
\title{summarize a bsseq dataset over defined regions (for e.g. cBioPortal)}
\usage{
atRegions(bsseq, regions, mappings = NULL, nm = "POETICname")
}
\arguments{
\item{bsseq}{a bsseq object}

\item{regions}{a GRanges or GRangesList of regions}

\item{mappings}{a mapping table, with rownames(mappings) == colnames(bsseq)}

\item{nm}{column of the mapping table to map to (default "POETICname")}
}
\description{
summarize a bsseq dataset over defined regions (for e.g. cBioPortal)
}
