% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigsToSegs.R
\name{bigWigsToSegs}
\alias{bigWigsToSegs}
\title{wrapper fn to process large numbers of bigWig'ed CN segmentations}
\usage{
bigWigsToSegs(suffix = ".CN.hg19.bw", filename = NULL)
}
\arguments{
\item{suffix}{suffix of segmentation files (default is ".CN.hg19.bw")}

\item{filename}{where to save the result (unsaved if NULL)}
}
\value{
\preformatted{     a data.frame just like the one from grToSeg 
}
}
\description{
wrapper fn to process large numbers of bigWig'ed CN segmentations
}
