% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuitMetadata.R
\name{biscuitMetadata}
\alias{biscuitMetadata}
\title{metadata about a Biscuit run using its vcfHeader metadata element}
\usage{
biscuitMetadata(bsseq)
}
\arguments{
\item{x}{a bsseq object with a $vcfHeader element}
}
\value{
information about the run
}
\description{
metadata about a Biscuit run using its vcfHeader metadata element
}
