% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{Choose loci or features by extremality}
\usage{
byExtremality(bsseq, r = NULL, k = 500)
}
\arguments{
\item{bsseq}{A bsseq object}

\item{r}{Regions to consider - NULL covers all loci (DEFAULT: NULL)}

\item{k}{How many rows/regions to return (DEFAULT: 500)}
}
\value{
\preformatted{  A matrix-like object with methylation values sorted by extremality
}
}
\description{
This function finds the k most extremal features (features above a certain
fraction of the Bernoulli variance) in 'bsseq' and returns their values.
}
\details{
For DNA methylation, particularly when summarized across regions, we can do
better (a lot better) than MAD. Since we know:
max(SD(X_j)) if X_j ~ Beta(a, b) < max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))
for X with a known mean and standard deviation (SD), then we can solve for
(a+b) by MoM. We can then define the extremality by:
extremality = sd(X_j) / bernoulliSD(mean(X_j))
}
\examples{

}
