% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBED.R
\name{checkBiscuitBED}
\alias{checkBiscuitBED}
\title{a BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample)}
\usage{
checkBiscuitBED(filename, sampleNames = NULL, merged = NULL)
}
\arguments{
\item{filename}{the file (compressed or not, doesn't matter) to load}

\item{sampleNames}{sample names for the bsseq object (if NULL, will create)}

\item{merged}{is the file a merged CpG file? (if NULL, will guess)}
}
\value{
\preformatted{       parameters for makeBSseq or makeBSseq_hdf5
}
}
\description{
a BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample)
}
\seealso{
load.biscuit
}
