% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clocks}
\alias{clocks}
\title{clocks}
\usage{
data(clocks, package="biscuiteer")
}
\description{
Epigenetic clock data
}
\details{
Source: See inst/scripts/clocks.R for how the clocks data object was
generated. For more information about sources, see the descriptions in
?getClock and ?WGBSage.
Return type: data.frame
}
\keyword{data}
