% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.biscuit.R
\name{combine.biscuit}
\alias{combine.biscuit}
\title{a bsseq combiner for Biscuit output (quite possibly sparse-matrix-backed)}
\usage{
combine.biscuit(x, y, sparse = FALSE)
}
\arguments{
\item{x}{a BSseq object}

\item{y}{another BSseq object}

\item{sparse}{should the combined object be sparse? (FALSE)}
}
\value{
\preformatted{       a combined bsseq::BSseq object
}
}
\description{
a bsseq combiner for Biscuit output (quite possibly sparse-matrix-backed)
}
\seealso{
BSseq

load.biscuit.merged

load.biscuit.unmerged
}
