% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condenseSampleNames.R
\name{condenseSampleNames}
\alias{condenseSampleNames}
\title{simplify sampleNames for a bsseq object}
\usage{
condenseSampleNames(tbx, stride, trailing = "\\\\.$")
}
\arguments{
\item{tbx}{a TabixFile instance to parse}

\item{stride}{how many columns per sample}

\item{trailing}{trailing character to trim ("\.$")}
}
\value{
a character vector of sample names (longest common substrings)
}
\description{
utility function for extracting samplenames from Tabix'ed sample columns
assumes VCF-style naming, such as with Sample_1.foo, Sample_1.bar,
or Sample1_foo,Sample1_bar, or really anything along these lines
}
