% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBsSeqCoverage.R
\name{correctBsSeqCoverage}
\alias{correctBsSeqCoverage}
\title{Tweaked version of TitanCNA's preprocessing}
\usage{
correctBsSeqCoverage(tumor, normal, bsgc = NULL, maps = NULL)
}
\arguments{
\item{tumor}{Binned tumor coverage - a GRanges from binCoverage}

\item{normal}{Binned normal coverage - a GRanges from binCoverage}

\item{bsgc}{Binned c2t GC content (DEFAULT: bsgc.hg19, 1 kb bins)}

\item{maps}{Binned c2t mappability (DEFAULT: maps.hg19, 1 kb bins)}
}
\value{
\preformatted{   Corrected tumor and normal read counts
}
}
\description{
Limits coverage to chromosomes 1 through 22 and X (for better or for worse)
}
\examples{

}
\seealso{
binCoverage

TitanCNA::correctReadDepth
}
