% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDepth.R
\name{correctDepth}
\alias{correctDepth}
\title{Tweaked version of HMMcopy's GC bias and mappability function}
\usage{
correctDepth(x, mappability = 0.9, samplesize = 50000,
  verbose = TRUE)
}
\arguments{
\item{x}{Binned hg19 coverage with GC content and mappability
columns}

\item{mappability}{Mappability threshold below which points are ignored
while creating the correction curve (value between 0
and 1)}

\item{samplesize}{Number of points sampled during LOESS fitting
(DEFAULT: 50 000)}

\item{verbose}{Print extra statements? (DEFAULT: TRUE)}
}
\value{
\preformatted{        Corrected tumor and normal read counts
}
}
\description{
Also used by correctBsSeqCoverage
}
\examples{

  bed <- system.file("extdata", "MCF7_Cunha_chr11p15.bed.gz",
                     package="biscuiteer")
  vcf <- system.file("extdata", "MCF7_Cunha_header_only.vcf.gz",
                     package="biscuiteer")
  bisc <- read.biscuit(BEDfile=bed, VCFfile=vcf, merged=FALSE)

  data(bsgc.hg19)
  bsgc <- bsgc.hg19
  data(maps.hg19)
  maps <- maps.hg19

  bisc <- binCoverage(bisc, bsgc)
  bsgc <- subsetByOverlaps(bsgc, bisc)
  maps <- subsetByOverlaps(maps, bisc)

  bisc$gc <- bsgc$score
  bisc$map <- maps$score
  names(mcols(bisc)) <- c("reads","gc","map")

  bisc_copy <- correctDepth(bisc)


}
\seealso{
correctBsSeqCoverage

HMMcopy::correctReadcount
}
