% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSR_subset.hg19}
\alias{ENSR_subset.hg19}
\title{ENSR_subset data from hg19 genome}
\usage{
data(ENSR_subset.hg19, package="biscuiteer")
}
\description{
Subset of ENSEMBL regulatory build regions for hg19 genome
}
\details{
Source URL: homo_sapiens.GRCh37.Regulatory_Build.regulatory_features.20161117.gff.gz
(regions that overlap Inifinium annotation manifests - described at
http://zwdzwd.github.io/InfiniumAnnotation - are selected for final GRanges)
Source type: GFF
Return type: GRanges
}
\keyword{data}
