% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSR_subset.hg38}
\alias{ENSR_subset.hg38}
\title{ENSR_subset data from hg38 genome}
\usage{
data(ENSR_subset.hg38, package="biscuiteer")
}
\description{
Subset of ENSEMBL regulatory build regions for hg19 genome
}
\details{
Source URL: homo_sapiens.GRCh38.Regulatory_Build.regulatory_features.20161111.gff.gz
(regions that overlap Inifinium annotation manifests - described at
http://zwdzwd.github.io/InfiniumAnnotation - are selected for final GRanges)
Source type: GFF
Return type: GRanges
}
\keyword{data}
