% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fexpit.R
\name{fexpit}
\alias{fexpit}
\title{Helper function: expanded expit}
\usage{
fexpit(x, sqz)
}
\arguments{
\item{x}{a vector of values between (-Inf, +Inf)}

\item{sqz}{the amount by which to 'squeeze', default is .000001}
}
\value{
\preformatted{   a vector of values in [0, 1]
}
}
\description{
Helper function: expanded expit
}
