% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLoci.R
\name{filterLoci}
\alias{filterLoci}
\title{Filter loci with zero coverage}
\usage{
filterLoci(bsseq, testCovariate)
}
\arguments{
\item{bsseq}{A bsseq object for filtering}

\item{testCovariate}{The name of the pData column dmrseq will test on}
}
\value{
\preformatted{          A bsseq object ready for dmrseq to use
}
}
\description{
Function potentially used to be a part of dmrseq. Included here to avoid
dmrseq failing due to any number of reasons related to lack of coverage.
}
\details{
The code is adapted from the precheck loop of dmrseq::dmrseq
}
\examples{

}
\seealso{
dmrseq

WGBSeq

RRBSeq
}
