% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixAge.R
\name{fixAge}
\alias{fixAge}
\title{project Horvath-type 'epigenetic clock' raw output onto actual ages}
\usage{
fixAge(x, adult = 21)
}
\arguments{
\item{x}{untransformed or raw prediction(s)}

\item{adult}{age of adulthood (default is 21)}
}
\value{
\preformatted{   transformed prediction(s) 
}
}
\description{
The 'Epigenetic Clock' (Horvath 2012) and similar schemes use a number of
CpG loci (or regions, or perhaps CpH loci -- it doesn't really matter what)
to estimate the chronological/biological age of samples from DNA methylation
with pre-trained feature weights (coefficients) for each region/locus.
}
\details{
All of these type of clocks use a nonlinear output transformation which
switches from an exponential growth model for children into a linear model
for adults, where \code{adult} is an arbitrary number (by default and custom,
that number is 21; elsewhere it can sometimes be seen as 20, but all known
epi-age transformation functions quietly add 1 to the constant internally).

This function implements the above standard output transformation step.
}
