% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuitMetadata.R
\name{getBiscuitMetadata}
\alias{getBiscuitMetadata}
\title{Alias for biscuitMetadata}
\usage{
getBiscuitMetadata(bsseq = NULL, VCF = NULL)
}
\description{
Alias for biscuitMetadata
}
\seealso{
biscuitMetadata
}
