% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClock.R
\name{getClock}
\alias{getClock}
\title{biscuiteer supports several "epigenetic clock" models. This fetches them.}
\usage{
getClock(model = c("horvath", "horvathshrunk", "hannum", "skinandblood"),
  padding = 15, genome = "hg19", useENSR = FALSE, useHMMI = FALSE)
}
\arguments{
\item{model}{one of "horvath", "horvathshrunk", "hannum", or "skinandblood"}

\item{padding}{how many base pairs (+/-) to expand a feature's footprint (15)}

\item{genome}{one of "hg19" (default), "GRCh37", "hg38", or "GRCh38" (hg19)}

\item{useENSR}{substitute ENSEMBL regulatory feature boundaries? (FALSE)}

\item{useHMMI}{substitute HMM-based CpG island boundaries? (FALSE)}
}
\value{
a List with elements \code{model}, \code{gr}, \code{intercept}, and \code{cleanup}
}
\description{
The remapped coordinates for the Horvath (2012) and Hannum (2013) clocks,
along with shrunken Horvath (2012) and improved Horvath (2018) models, are
provided as part of biscuiteer (visit inst/extdata/clocks.R to find out how)
along with some functionality to make them more usable in RRBS/WGBS data of
varying coverage along varying genomes. For example, the HMM-based CpG island
model introduced by Wu (2010) can be used to assign to within-island features
the methylation rate of their associated island, and ENSEMBL regulatory build
features (ENSR features, for short) such as CTCF binding sites can have their
coordinates substituted for the default padded boundaries of a feature.
}
\details{
The net result of this process is that, while the default settings simply
swap in a 30-bp stretch centered on the selected clock's CpG (and/or CpH)
loci, add the intercept, and ship out the model, much more flexibility is
available to the user. This function provides a single point for tuning of
such options in the event that defaults don't work well for a user.

The precedence of options is as follows:
\itemize{
\item if a feature has neither ENSR nor HMMI IDs, it is padded (only) +/- bp.
\item if it has an HMMI but not ENSR ID or ENSR==FALSE, the HMM island is used.
\item if a feature has an ENSR ID, and ENSR==TRUE, the ENSR feature is used.
}

If a feature has both an ENSR ID and an HMMI ID, and both options are TRUE,
then the ENSR start and end coordinates will take precedence over its HMMI.

The above shenanigans produce the GRanges object returned as \code{gr} in a List.
The \code{intercept} value returned with the model is its fixed (B0) coefficient.
The \code{cleanup} function returned with the model transforms its raw output.
}
