% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogitFracMeth.R
\name{getLogitFracMeth}
\alias{getLogitFracMeth}
\alias{getMvals}
\title{Helper function for compartment inference}
\usage{
getLogitFracMeth(x, minCov = 3, minSamp = 2, k = 0.1, r = NULL)

getMvals(x, minCov = 3, minSamp = 2, k = 0.1, r = NULL)
}
\arguments{
\item{x}{A bsseq object with methylated and total reads}

\item{minCov}{Minimum read coverage for landmarking samples (DEFAULT: 3)}

\item{minSamp}{Minimum landmark samples with at least minCov (DEFAULT: 2)}

\item{k}{Pseudoreads for smoothing (DEFAULT: 0.1)}

\item{r}{Regions to collapse over - if NULL, do it by CpG
(DEFAULT: NULL)}
}
\value{
\preformatted{    Smoothed logit(M / Cov) matrix with coordinates as row names
}
}
\description{
Want an object with nominally Gaussian error for compartment inference, so
this function uses 'suitable' (defaults to to 3 or more reads in 2 or more
samples) measurements. Using Dirichlet smoothing (adding 'k' reads to M
and U), these measurements are then turned into lightly moderated,
logit-transformed methylated-fraction estimates for compartment calling.
}
\section{Functions}{
\itemize{
\item \code{getMvals}: Alias for getLogitFracMeth
}}

\examples{

  tcga_bed <- system.file("extdata", "TCGA_BLCA_A13J_chr11p15_merged.bed.gz",
                          package = "biscuiteer")
  tcga_vcf <- system.file("extdata", "TCGA_BLCA_A13J_header_only.vcf.gz",
                          package = "biscuiteer")
  bisc <- read.biscuit(BEDfile = tcga_bed, VCFfile = tcga_vcf,
                       merged = TRUE, genome = "hg38")

  reg <- GRanges(seqnames = rep("chr11",5),
                 strand = rep("*",5),
                 ranges = IRanges(start = c(0,2.8e6,1.17e7,1.38e7,1.69e7),
                                  end= c(2.8e6,1.17e7,1.38e7,1.69e7,2.2e7))
                 )

  getLogitFracMeth(bisc, minSamp = 1, r = reg)

}
