% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMvals.R
\name{getMvals}
\alias{getMvals}
\title{helper function for compartment inference (shrink-by-smoothing)}
\usage{
getMvals(x, minCov = 3, minSamp = 2, k = 1)
}
\arguments{
\item{x}{a BSseq object with methylated and total reads}

\item{minCov}{minimum read coverage for landmarking samples (3)}

\item{minSamp}{minimum landmark samples with >= minCov (2)}

\item{k}{pseudoreads for smoothing (1)}
}
\value{
\preformatted{       a matrix of Mvalues, with genomic coordinates as row names
}
}
\description{
since we want something resembling M-values for compartment inference,
this function grabs suitably deep (>= 3 reads in >=2 sample) measurements
and turns them into lightly moderated M-values for compartment calling
by Dirichlet smoothing (adding addReads reads to each of M and Covg)
}
