% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg19.chromArm}
\alias{hg19.chromArm}
\title{hg19.chromArm}
\usage{
data(hg19.chromArm, package="biscuiteer")
}
\description{
Chromosome arm locations for hg19 genome
}
\details{
Source URL: http://hgdownload.cse.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz
(Chromosome arms were combined to form the final GRanges)
Source type: TXT
Return type: GRanges
}
\keyword{data}
