% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMM_CpG_islands.hg19}
\alias{HMM_CpG_islands.hg19}
\title{HMM_CpG_islands.hg19}
\usage{
data(HMM_CpG_islands.hg19, package="biscuiteer")
}
\description{
Hidden Markov Model-derived CpG islands from hg19 genome
}
\keyword{data}
