% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HMM_CpG_islands.hg38}
\alias{HMM_CpG_islands.hg38}
\title{HMM_CpG_islands.hg38}
\usage{
data(HMM_CpG_islands.hg38, package="biscuiteer")
}
\description{
Hidden Markov Model-derived CpG islands from hg38 genome
}
\details{
Source URL: https://www.ncbi.nlm.nih.gov/pubmed/20212320
(Hidden Markov Model CpG islands were produced using the method described in
this paper. The hg19 genome was used for the CpG island production.)
Source type: hg19 genome and procedure described in paper
Return type: GRanges
}
\keyword{data}
