% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.biscuit.merged.R
\name{load.biscuit.merged}
\alias{load.biscuit.merged}
\title{a bsseq loader for Biscuit mergecg output (BED-like format, 3 cols/sample)
e.g. P01-028-T06.joint.cg.merged.hg19.bed.gz has 3 samples, and thus 12 cols}
\usage{
load.biscuit.merged(params)
}
\arguments{
\item{params}{parameters from having run checkBiscuitBED}
}
\value{
\preformatted{       a BSseq object from the bsseq package
}
}
\description{
a bsseq loader for Biscuit mergecg output (BED-like format, 3 cols/sample)
e.g. P01-028-T06.joint.cg.merged.hg19.bed.gz has 3 samples, and thus 12 cols
}
\seealso{
load.biscuit.unmerged
}
