% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.biscuit.R
\name{load.biscuit}
\alias{load.biscuit}
\title{a bsseq loader for Biscuit output (BED-like format, 2 or 3 cols/sample)
e.g. P01-028-T06.joint.ch.hg19.bed.gz has 3 samples, and 9 columns total,
while P01-028-T06.joint.cg.merged.hg19.bed.gz has 3 samples and 12 columns}
\usage{
load.biscuit(filename, sampleNames = NULL, hdf5 = FALSE, merged = NULL)
}
\arguments{
\item{filename}{the file (compressed or not, doesn't matter) to load}

\item{sampleNames}{sample names for the bsseq object (if NULL, will create)}

\item{hdf5}{make the object HDF5-backed? (FALSE)}

\item{merged}{is the file a merged CpG file? (if NULL, will guess)}
}
\value{
\preformatted{       a bsseq::BSseq object, possibly HDF5-backed
}
}
\description{
a bsseq loader for Biscuit output (BED-like format, 2 or 3 cols/sample)
e.g. P01-028-T06.joint.ch.hg19.bed.gz has 3 samples, and 9 columns total,
while P01-028-T06.joint.cg.merged.hg19.bed.gz has 3 samples and 12 columns
}
\seealso{
BSseq

checkBiscuitBED

load.biscuit.merged

load.biscuit.unmerged
}
