% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHDF5.R
\name{loadHDF5}
\alias{loadHDF5}
\title{Function to load HDF5-backed SummarizedExperiment from read.biscuit}
\usage{
loadHDF5(dir, prefix = "")
}
\arguments{
\item{dir}{path to the HDF5-backed file}

\item{prefix}{optional prefix to add to filenames inside ‘dir’; default is ""}
}
\value{
\preformatted{   a SummarizedExperiment-derived object (such as a BSseq object)
}
}
\description{
Local wrapper for HDF5Array::loadHDF5SummarizedExperiment
}
