% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{make an in-core BSseq object from the results of reading a Biscuit BED file}
\usage{
makeBSseq(tbl, betacols, covgcols, pData, sparse = TRUE)
}
\arguments{
\item{tbl}{a tibble (from read_tsv)}

\item{betacols}{the beta column names (from checkBiscuitBED)}

\item{covgcols}{the coverage column names (from checkBiscuitBED)}

\item{pData}{a DataFrame (usually from checkBiscuitBED)}

\item{sparse}{make the object Matrix-backed? (TRUE)}
}
\value{
an in-core BSseq object
}
\description{
make an in-core BSseq object from the results of reading a Biscuit BED file
}
\seealso{
makeBSseq_HDF5
}
