% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq_hdf5.R
\name{makeBSseq_hdf5}
\alias{makeBSseq_hdf5}
\title{make a HDF5-backed BSseq object from a biscuit BED}
\usage{
makeBSseq_hdf5(tbl, params, grid, M_sink, Cov_sink, sink_lock,
  simplify = FALSE, verbose = FALSE)
}
\arguments{
\item{tbl}{a tibble (from read_tsv) or a data.table (from fread)}

\item{params}{parameters from checkBiscuitBED}

\item{grid}{RegularArrayGrid for writing HDF5 file}

\item{M_sink}{HDF5RealizationSink for M matrix}

\item{Cov_sink}{HDF5RealizationSink for Cov matrix}

\item{sink_lock}{IPC lock to avoid race condition when writing to HDF5 file}

\item{simplify}{simplify sample names by dropping .foo.bar.hg19 and similar (FALSE)}

\item{verbose}{be verbose about what is happening (FALSE)}
}
\value{
\preformatted{     a HDF5-backed BSseq object
}
}
\description{
make a HDF5-backed BSseq object from a biscuit BED
}
\seealso{
makeBSseq
}
