% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNAs.R
\name{plotCNA}
\alias{plotCNA}
\title{Analyze CNAs}
\usage{
plotCNA(filenames, minWidth = 1000)
}
\arguments{
\item{filenames}{List of files with tumor and normal Excel sheets}

\item{minWidth}{Scaling factor (DEFAULT: 1000)}
}
\value{
\preformatted{      Heatmap of CNAs
}
}
\description{
Analyze CNAs
}
