% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistate.R
\name{plotEpiread}
\alias{plotEpiread}
\title{Plot the results of tabulateEpibed() as a
quasi-lollipop plot.}
\usage{
plotEpiread(mat)
}
\arguments{
\item{mat}{Input matrix that comes out of tabulateEpibed()}
}
\value{
An epiread plot
}
\description{
Plot the results of tabulateEpibed() as a
quasi-lollipop plot.
}
\examples{
epibed.nome <- system.file("extdata", "hct116.nome.epiread",
                           package="biscuiteer")
epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
                             genome = "hg19")
epibed.tab.nome <- tabulateEpibed(epibed.nome.gr)
plotEpiread(epibed.tab.nome$gc_table)
}
