% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegs.R
\name{plotSegs}
\alias{plotSegs}
\title{Plot WGBScopy results}
\usage{
plotSegs(cn, segs = NULL, chr = "chr22", name = "Segmented", ...)
}
\arguments{
\item{cn}{Binned, corrected read depths from correctBsSeqCoverage}

\item{segs}{Segmentation result from WGBScopy}

\item{chr}{Chromosome to display (DEFAULT: "chr22")}

\item{name}{Name for the plot (DEFAULT: "Segmented")}

\item{...}{Other arguments to pass on to plot}
}
\description{
Modified from HMMcopy
}
\examples{

}
