% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{plotVafAndCN}
\alias{plotVafAndCN}
\title{Plot variant annotation matrix and CN}
\usage{
plotVafAndCN(VRL, rse, cutoff = 0.25, minCN = 0.25, minWidth = 1000)
}
\arguments{
\item{VRL}{VRangesList of variants}

\item{rse}{RaggedExperiment of CNs}

\item{cutoff}{Minimum value to plot for variants (DEFAULT: 0.25)}

\item{minCN}{Minimum CN to plot (DEFAULT: 0.25)}

\item{minWidth}{Scaling factor for CNs (DEFAULT: 1000)}
}
\value{
\preformatted{     A Heatmap of the variants and CNs
}
}
\description{
Plot variant annotation matrix and CN
}
