% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{plotVafMat}
\alias{plotVafMat}
\title{Plot variant annotation matrix}
\usage{
plotVafMat(VRL, cutoff = 0.25, ANNonly = FALSE)
}
\arguments{
\item{VRL}{VRangesList of variants}

\item{cutoff}{Minimum value to plot (DEFAULT: 0.25)}

\item{ANNonly}{Show annotations only (DEFAULT: FALSE)}
}
\value{
\preformatted{    A Heatmap of the variants
}
}
\description{
Plot variant annotation matrix
}
