% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.segs.R
\name{read.segs}
\alias{read.segs}
\title{Function to turn segmentations into GRanges with correction(s)}
\usage{
read.segs(file, weight = TRUE, by = c("marks", "width"),
  from = c(0.4, 0.6), minNorm = -0.5, maxNorm = +0.5)
}
\arguments{
\item{file}{The segmentation file to process}

\item{weight}{Add weights for per-subject baselining? (DEFAULT: TRUE)}

\item{by}{If weighting, do so by marks or by width (DEFAULT: "marks")}

\item{from}{Quantiles around the median to "squash" to baseline
(DEFAULT: c(.4, .6))}

\item{minNorm}{Min log2ratio to consider as possible "normal" baseline
(DEFAULT: -0.5)}

\item{maxNorm}{Max log2ratio to consider as possible "normal" baseline
(DEFAULT: +0.5)}
}
\value{
\preformatted{    A GRanges with corrections applied
}
}
\description{
If your segmentation file is not vaild, this probably won't work. A smarter
way to squash baselines would be with a mixture model. One bonus is that the
baselining is self-correcting - applying it to an already-corrected
segmentation will not do any harm and may even help.
}
