% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpibed.R
\name{removeInsertions}
\alias{removeInsertions}
\title{Function to remove insertions to reset genomic coordinates
for plotting of read or fragment level methylation states,
relative to the reference genome.}
\usage{
removeInsertions(gr, is.nome = FALSE)
}
\arguments{
\item{gr}{Input is a read or fragment level GRanges out of readEpibed}

\item{is.nome}{Whether the input is NOMe-seq or not}
}
\value{
A new GRanges object with reset genomic coordinates and filtered
CG and/or GC RLE strings
}
\description{
Function to remove insertions to reset genomic coordinates
for plotting of read or fragment level methylation states,
relative to the reference genome.
}
\examples{

epibed.nome <- system.file("extdata", "hct116.nome.epiread.gz",
                           package="biscuiteer")
epibed.bsseq <- system.file("extdata", "hct116.bsseq.epiread.gz",
                            package="biscuiteer")
epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
                             genome = "hg19", chr = "chr1")
epibed.bsseq.gr <- readEpibed(epibed = epibed.bsseq,
                              genome = "hg19", chr = "chr1")
                              
epibed.nome.filt.gr <- removeInsertions(epibed.nome.gr, is.nome = TRUE)
epibed.bsseq.filt.gr <- removeInsertions(epibed.bsseq.gr)
}
