% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segToGr.R
\name{segToGr}
\alias{segToGr}
\title{reverse a grToSeg... import a segmentation file}
\usage{
segToGr(seg, genome = "hg19", name = "ID", score = "seg.mean")
}
\arguments{
\item{seg}{the .seg filename}

\item{genome}{genome against which segments were annotated ("hg19")}

\item{name}{.seg file column to use as $name metadata ("ID")}

\item{score}{.seg file column to use as $score metadata ("seg.mean")}
}
\value{
\preformatted{     a GRanges
}
}
\description{
reverse a grToSeg... import a segmentation file
}
