% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqinfo.mm10}
\alias{seqinfo.mm10}
\title{seqinfo.mm10}
\usage{
data(seqinfo.mm10, package="biscuiteer")
}
\description{
Seqinfo for mm10 genome
}
\details{
Source URL: http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips/mm10.chrom.sizes
(The output from this site was downloaded into a TXT file and then loaded
into a sorted Seqinfo table)
Source type: TXT
Return type: Seqinfo
}
\keyword{data}
