% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifySampleNames.R
\name{simplifySampleNames}
\alias{simplifySampleNames}
\title{Simplify bsseq sample names (usually VCF columns)}
\usage{
simplifySampleNames(x)
}
\arguments{
\item{x}{a BSseq (or any SummarizedExperiment-derived) object}
}
\value{
\preformatted{   the same object but with simplified sample names
}
}
\description{
Tries using the longest common subsequence to figure out what can be dropped.
}
