% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBsSeqOver.R
\name{summarizeBsSeqOver}
\alias{summarizeBsSeqOver}
\title{Summarize methylation over provided regions}
\usage{
summarizeBsSeqOver(bsseq, segs, dropNA = FALSE, impute = FALSE)
}
\arguments{
\item{bsseq}{The bsseq object to summarize}

\item{segs}{Regions to summarize over (GRanges object, no GRangesList yet)}

\item{dropNA}{Whether to drop rows if more than half of samples are NA
(DEFAULT: FALSE)}

\item{impute}{Whether to impute NAs/NaNs (DEFAULT: FALSE)}
}
\value{
\preformatted{   A matrix of regional methylation fractions
}
}
\description{
Used for bsseq objects. Mostly a local wrapp for getMeth.
}
\examples{

}
