% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixRetrieve}
\alias{tabixRetrieve}
\title{Read from tabix-indexed bed file to list objects}
\usage{
tabixRetrieve(paths, chrm, beg = 1, end = 2^28, min_depth = 0,
  sample_names = NULL, BPPARAM = SerialParam())
}
\arguments{
\item{paths}{paths to the bed files}

\item{chrm}{chromosome name}

\item{beg}{start coordinate of CpG}

\item{end}{end coordinate of CpG 2^29 is the max taken}

\item{sample_names}{sample names, just use paths if not specified}

\item{BPPARAM}{how to parallelize}
}
\value{
a list object with DNA methylation level and depth
}
\description{
Read from tabix-indexed bed file to list objects
}
