% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixToGr}
\alias{tabixToGr}
\title{Read from tabix-indexed bed file to GRanges or tibble
The fourth column is DNA methylation level and the fifth is depth.
This is also the default output of Biscuit BED.}
\usage{
tabixToGr(paths, chrm, beg, end, as_gr = FALSE, min_depth = 0,
  sample_names = NULL, BPPARAM = SerialParam())
}
\arguments{
\item{paths}{paths to the bed files}

\item{chrm}{chromosome name}

\item{beg}{start coordinate of CpG}

\item{end}{end coordinate of CpG}

\item{as_gr}{whether to output GRanges}

\item{min_depth}{mask all depth < min_depth as NA. This is useful as depths
are not recorded in the output}

\item{sample_names}{sample names, just use paths if not specified}

\item{BPPARAM}{how to parallelize}
}
\value{
a tibble or GRanges object with DNA methylation level
}
\description{
Read from tabix-indexed bed file to GRanges or tibble
The fourth column is DNA methylation level and the fifth is depth.
This is also the default output of Biscuit BED.
}
