% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixToGRanges}
\alias{tabixToGRanges}
\title{Read from tabix-indexed bed file into a GRanges}
\usage{
tabixToGRanges(..., min_depth = 0)
}
\arguments{
\item{...}{tabixRetrieve arguments}

\item{min_depth}{minimum depth as depth is discarded in GRanges}
}
\value{
a GenomicRanges of DNA methylation level
}
\description{
Read from tabix-indexed bed file into a GRanges
}
\seealso{
tabixRetrieve
}
