% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixToMatrix}
\alias{tabixToMatrix}
\title{Read from tabix-indexed bed file into a matrix
The row names are the CpG coordinates. The column names are
sample names.}
\usage{
tabixToMatrix(..., min_depth = 0)
}
\arguments{
\item{...}{tabixRetrieve and tabixToTable arguments}

\item{min_depth}{minimum depth as depth is discarded in GRanges}
}
\value{
a matrix of DNA methylation level
}
\description{
Read from tabix-indexed bed file into a matrix
The row names are the CpG coordinates. The column names are
sample names.
}
\seealso{
tabixRetrieve

tabixToTable
}
