% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixToTable}
\alias{tabixToTable}
\title{Read from tabix-indexed bed file into a tibble}
\usage{
tabixToTable(..., min_depth = 0)
}
\arguments{
\item{...}{tabixRetrieve arguments}

\item{min_depth}{minimum depth as depth is discarded in GRanges}
}
\value{
a tibble of DNA methylation level
}
\description{
Read from tabix-indexed bed file into a tibble
}
\seealso{
tabixRetrieve
}
