% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSage.R
\name{WGBSage}
\alias{WGBSage}
\title{guess a specimen's age from a modified Horvath algorithm}
\usage{
WGBSage(x)
}
\arguments{
\item{x}{a BSseq object}
}
\value{
age   age estimate for each column of x
}
\description{
guess a specimen's age from a modified Horvath algorithm
}
