% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSseg.R
\name{WGBSseg}
\alias{WGBSseg}
\title{Copy number segmentation of GC- and mapping-corrected WGBS coverage}
\usage{
WGBSseg(cn, minAbs = 0.3, minWidth = 20000, ...)
}
\arguments{
\item{cn}{Binned, corrected read depths from correctBsSeqCoverage}

\item{minAbs}{Minimum absolute value of a segment to keep (DEFAULT: 0.3)}

\item{minWidth}{Minimum width of segment (NOT A GAP) to keep
(DEFAULT: 20 000)}

\item{...}{Other parameters passed to fastseg}
}
\value{
\preformatted{     A GRanges (hg19 seqinfo)
}
}
\description{
Copy number segmentation of GC- and mapping-corrected WGBS coverage
}
\examples{

  library(fastseg)
  library(HMMcopy) 
  library(rtracklayer)
  files <- list.files(patt="P01-022-.*covg.hg19.bw") 
  names(files) <- sapply(strsplit(files, "(\\\\.|\\\\-)"), `[`, 3)
  covgs <- sapply(files, import)

  RMS1 <- correctBsSeqCoverage(covgs$tumor1, covgs$normal)
  RMS1seg <- WGBSseg(RMS1)
  plotSegs(RMS1, RMS1seg, "chr10")

  RMS2 <- correctBsSeqCoverage(covgs$tumor2, covgs$normal)
  RMS2seg <- WGBSseg(RMS2) 
  plotSegs(RMS2, RMS2seg, "chr10")

}
