% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Zhou_solo_WCGW_inCommonPMDs.hg19}
\alias{Zhou_solo_WCGW_inCommonPMDs.hg19}
\title{Zhou_solo_WCGW_inCommonPMDs.hg19}
\usage{
data(Zhou_solo_WCGW_inCommonPMDs.hg19, package="biscuiteer")
}
\description{
Dataset description
}
\keyword{data}
