% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atRegions.R
\name{atRegions}
\alias{atRegions}
\title{Summarize a bsseq dataset over defined regions}
\usage{
atRegions(bsseq, regions, mappings = NULL, nm = "POETICname", ...)
}
\arguments{
\item{bsseq}{A bsseq object}

\item{regions}{A GRanges or GRangesList of regions}

\item{mappings}{A mapping table with rownames(mappings) == colnames(bsseq)
(DEFAULT: NULL)}

\item{nm}{Column of the mapping table to map to
(DEFAULT: "POETICname")}

\item{...}{Other arguments to pass to summarizeBsSeqOver}
}
\value{
\preformatted{     Summarized information about the bsseq object for the given
               DNA regions
}
}
\description{
Calls summarizeBsSeqOver to summarize a bsseq object over provided DNA
regions. Useful for exploring genomic data using cBioPortal.
}
\examples{

}
