% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigWigsToSegs.R
\name{bigWigsToSegs}
\alias{bigWigsToSegs}
\title{Process many bigWig'ed CN segmentations}
\usage{
bigWigsToSegs(bwList, suffix = ".CN.hg19.bw", filename = NULL)
}
\arguments{
\item{bwList}{bigWig files to read in}

\item{suffix}{Suffix of segmentation files (DEFAULT: ".CN.hg19.bw")}

\item{filename}{Where to save the result - unsaved if NULL (DEFAULT: NULL)}
}
\value{
\preformatted{     A data.frame like the one returned from grToSeg
}
}
\description{
Wrapper function to grToSeg to make processing large numbers of bigWigs easier
}
\examples{

}
