% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCoverage.R
\name{binCoverage}
\alias{binCoverage}
\title{Bin CpG or CpH coverage to simplify and improve CNA "sketching"}
\usage{
binCoverage(bsseq, bins, which = NULL, QDNAseq = TRUE, readLen = 100)
}
\arguments{
\item{bsseq}{A bsseq object - supplied to getCoverage()}

\item{bins}{Bins to summarize over - from tileGenome or QDNAseq.xxYY}

\item{which}{Limit to specific regions? - functions as an import()
(DEFAULT: NULL)}

\item{QDNAseq}{Return a QDNAseqReadCounts? - if FALSE, returns a GRanges
(DEFAULT: TRUE)}

\item{readLen}{Correction factor for coverage - read length in bp
(DEFAULT: 100)}
}
\value{
\preformatted{    Binned read counts
}
}
\description{
Example usage for E-M
}
\examples{

}
