% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCoverage.R
\name{binCoverage}
\alias{binCoverage}
\title{bin CpG/CpH coverage to simplify & improve CNA "sketching" (e.g. for E-M)}
\usage{
binCoverage(x, bins, which = NULL, QDNAseq = TRUE)
}
\arguments{
\item{x}{a bsseq object, maybe HDF5-backed, for getCoverage()}

\item{bins}{bins to summarize over (from tileGenome or QDNAseq.xxYY)}

\item{which}{limit to specific regions? (functions as in import())}

\item{QDNAseq}{return a QDNAseqReadCounts? (if FALSE, return a GRanges)}
}
\value{
\preformatted{     binned read counts
}
}
\description{
FIXME: turn this into a generic for bsseq objects, for bigWigs, for \link{...}
FIXME: figure out how to estimate the most likely GCbias ~ DNAme linkage
}
