% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binCoverage.R
\name{binCoverage}
\alias{binCoverage}
\title{helper function for binning .covg.hg19.bw bigWigs (see below for note)
nb. it is best to extract these from Biscuit BAMs via genomeCoverageBed
FIXME: figure out how to estimate the most likely GCbias ~ DNAme linkage}
\usage{
binCoverage(covg, bins = NULL, tilewidth = 1000, which = NULL)
}
\arguments{
\item{covg}{the coverage track (a GRanges) or a bigWig filename}

\item{bins}{the bins to summarize over (use tileGenome to generate)}

\item{tilewidth}{if is.null(bins), what width should be used? (1000bp)}

\item{which}{if covg is a filename, what regions should be imported?}
}
\value{
\preformatted{     binned read counts
}
}
\description{
helper function for binning .covg.hg19.bw bigWigs (see below for note)
nb. it is best to extract these from Biscuit BAMs via genomeCoverageBed
FIXME: figure out how to estimate the most likely GCbias ~ DNAme linkage
}
