% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuitMetadata.R
\name{biscuitMetadata}
\alias{biscuitMetadata}
\alias{getBiscuitMetadata}
\title{metadata about a Biscuit run using its vcfHeader metadata element, or its VCF}
\usage{
biscuitMetadata(x = NULL, VCF = NULL)
}
\arguments{
\item{x}{a bsseq object with a $vcfHeader element (or NULL)}

\item{VCF}{a tabix'ed VCF file (can just be the header) whence x is derived}
}
\value{
\preformatted{ information about the run
}
}
\description{
metadata about a Biscuit run using its vcfHeader metadata element, or its VCF
}
