% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuitMetadata.R
\name{biscuitMetadata}
\alias{biscuitMetadata}
\alias{getBiscuitMetadata}
\title{Biscuit metadata from VCF header}
\usage{
biscuitMetadata(bsseq = NULL, VCF = NULL)

getBiscuitMetadata(bsseq = NULL, VCF = NULL)
}
\arguments{
\item{bsseq}{A bsseq object with a vcfHeader element (DEFAULT: NULL)}

\item{VCF}{A tabix'ed VCF file (can just be the header information)
from which the bsseq vcfHeader element is derived
(DEFAULT: NULL)}
}
\value{
\preformatted{  Information regarding the Biscuit run
}
}
\description{
Returns metadata from a Biscuit run using either a supplied VCF file or
the vcfHeader metadata element from the bsseq object
}
\section{Functions}{
\itemize{
\item \code{getBiscuitMetadata}: Alias for biscuitMetadata
}}

\examples{

  tcga_bed <- system.file("extdata", "TCGA_BLCA_A13J_chr11p15_merged.bed.gz",
                          package = "biscuiteer")
  tcga_vcf <- system.file("extdata", "TCGA_BLCA_A13J_header_only.vcf.gz",
                          package = "biscuiteer")
  bisc <- read.biscuit(BEDfile = tcga_bed, VCFfile = tcga_vcf,
                       merged = TRUE, genome = "hg38", verbose = TRUE)

  meta <- biscuitMetadata(bisc)
  meta
  
}
