% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biscuitMetadata.R
\name{biscuitMetadata}
\alias{biscuitMetadata}
\alias{getBiscuitMetadata}
\title{Biscuit metadata from VCF header}
\usage{
biscuitMetadata(bsseq = NULL, VCF = NULL)
}
\arguments{
\item{bsseq}{A bsseq object with a vcfHeader element (DEFAULT: NULL)}

\item{VCF}{A tabix'ed VCF file (can just be the header information)
from which the bsseq vcfHeader element is derived
(DEFAULT: NULL)}
}
\value{
\preformatted{  Information regarding the Biscuit run
}
}
\description{
Returns metadata from a Biscuit run using either a supplied VCF file or
the vcfHeader metadata element from the bsseq object
}
\examples{

}
