% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byChromArm.R
\name{byChromArm}
\alias{byChromArm}
\title{an idiotically simple parallelization step that I've been using for years}
\usage{
byChromArm(x, arms = NULL)
}
\arguments{
\item{x}{anything with a GRanges in it: BSseq, SummarizedExperiment, etc.}

\item{arms}{(optional) another GRanges, but just specifying chromosome arms.}
}
\value{
\preformatted{ a list, List, or *List, with pieces of x by chromosome arm.
}
}
\description{
This function just splits an object by chromosome arm, which tends to make
parallelization MUCH easier, as cross-arm dependencies are unusual, and the
larger chromosomes can thus be split across processes or machines without
worrying much about data starvation for processes on smaller chromosomes.
}
