% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byChromArm.R
\name{byChromArm}
\alias{byChromArm}
\title{A simple parallization step}
\usage{
byChromArm(x, arms = NULL)
}
\arguments{
\item{x}{Any object with a GRanges in it: bsseq, SummarizedExperiment...}

\item{arms}{Another GRanges, but specifying chromosome arms (DEFAULT: NULL)}
}
\value{
\preformatted{ A list, List, or *list, with pieces of x by chromosome arm
}
}
\description{
This function splits an object by chromosome arm, which tends to make
parallelization much easier, as cross-arm dependencies are unusual.
Therefore, the larger chromosomes can be split across processes or machines
without worrying much about data starvation for processes on smaller
chromosomes.
}
\examples{

  tcga_bed <- system.file("extdata", "TCGA_BLCA_A13J_chr11p15_merged.bed.gz",
                          package = "biscuiteer")
  tcga_vcf <- system.file("extdata", "TCGA_BLCA_A13J_header_only.vcf.gz",
                          package = "biscuiteer")
  bisc <- read.biscuit(BEDfile = tcga_bed, VCFfile = tcga_vcf,
                       merged = TRUE, genome = "hg38")

  reg <- GRanges(seqnames = rep("chr11",5),
                 strand = rep("*",5),
                 ranges = IRanges(start = c(0,2.8e6,1.17e7,1.38e7,1.69e7),
                                  end= c(2.8e6,1.17e7,1.38e7,1.69e7,2.2e7))
                 )
  names(reg) <- as.character(reg)

  arms <- byChromArm(bisc, arms = reg)
  lapply(arms, rowRanges)

}
