% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byChromArm.R
\name{byChromArm}
\alias{byChromArm}
\title{A simple parallization step}
\usage{
byChromArm(x, arms = NULL)
}
\arguments{
\item{x}{Any object with a GRanges in it: bsseq, SummarizedExperiment...}

\item{arms}{Another GRanges, but specifying chromosome arms (DEFAULT: NULL)}
}
\value{
\preformatted{ A list, List, or *list, with pieces of x by chromosome arm
}
}
\description{
This function splits an object by chromosome arm, which tends to make
parallelization much easier, as cross-arm dependencies are unusual.
Therefore, the larger chromosomes can be split across processes or machines
without worrying much about data starvation for processes on smaller
chromosomes.
}
\examples{

}
