% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{choose loci or features by extremality (fraction of Bernoulli variance)}
\usage{
byExtremality(x, r = NULL, k = 500)
}
\arguments{
\item{x}{a bsseq object}

\item{r}{regions to consider (default is NULL, all loci... !)}

\item{k}{how many rows/regions to return (500)}
}
\value{
a matrix-like object with methylation values, by extremality
}
\description{
For DNA methylation, particularly summarized across regions,
we can do better (a lot better) than MAD.  Since we know that
}
\details{
max(SD(X_j)) if X_j ~ Beta(a, b) < max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))

for X having a known mean and SD, hence solvable for a + b by MoM, define

extremality = sd(X_j) / bernoulliSD(mean(X_j))

This function finds the k most extremal features in x & returns their values.
For count-based proportions,
}
