% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBiscuitBED.R
\name{checkBiscuitBED}
\alias{checkBiscuitBED}
\title{A BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample).
By default, files with over 50 million loci will be processed iteratively,
since data.table tends to run into problems with .gzipped joint CpH files.
This function absolutely assumes that BED files are tabixed. No exceptions!}
\usage{
checkBiscuitBED(filename, sampleNames = NULL, yieldSize = 50000000,
  merged = NULL)
}
\arguments{
\item{filename}{the file (compressed and tabixed, with header) to load}

\item{sampleNames}{sample names (if NULL, create; if data.frame, make pData)}

\item{yieldSize}{for files > \code{yieldSize} lines long, chunk the file (5e7)}
}
\value{
\preformatted{       parameters for makeBSseq or makeBSseq_hdf5
}
}
\description{
A BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample).
By default, files with over 50 million loci will be processed iteratively,
since data.table tends to run into problems with .gzipped joint CpH files.
This function absolutely assumes that BED files are tabixed. No exceptions!
}
\seealso{
\preformatted{      load.biscuit
}
}
