% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBiscuitBED.R
\name{checkBiscuitBED}
\alias{checkBiscuitBED}
\title{A BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample).
By default, files with over 50 million loci will be processed iteratively,
since data.table tends to run into problems with .gzipped joint CpH files.
This function absolutely assumes that BED files are tabixed. No exceptions!}
\usage{
checkBiscuitBED(BEDfile, VCFfile = NULL, sampleNames = NULL,
  chunkSize = 50000000, merged = NULL, hdf5 = FALSE, sparse = TRUE,
  how = c("data.table", "readr"), chr = NULL)
}
\arguments{
\item{BEDfile}{a BED-like file (compressed and tabixed, maybe w/header)}

\item{VCFfile}{a VCF file (compressed and tabixed; only needs the header)}

\item{sampleNames}{if NULL create; if vector assign; if data.frame make pData}

\item{chunkSize}{for files > \code{yieldSize} lines long, chunk the file (5e7)}

\item{merged}{boolean; is this merged CpG data? (NULL; guess if merged)}

\item{hdf5}{boolean; use HDF5 arrays for backing of the data? (FALSE)}

\item{sparse}{boolean; use sparse Matrix objects for the data? (TRUE)}

\item{how}{how to load the data? "data.table" (default) or "readr"}

\item{chr}{load a specific chromosome (to rbind() later)? (NULL)}
}
\value{
\preformatted{       parameters for makeBSseq or makeBSseq_hdf5
}
}
\description{
A BED checker for Biscuit cg/ch output (BED-like format, 2 or 3 cols/sample).
By default, files with over 50 million loci will be processed iteratively,
since data.table tends to run into problems with .gzipped joint CpH files.
This function absolutely assumes that BED files are tabixed. No exceptions!
}
\seealso{
\preformatted{      load.biscuit
}
}
