% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBsSeqCoverage.R
\name{correctBsSeqCoverage}
\alias{correctBsSeqCoverage}
\title{Tweaked version of TitanCNA's preprocessing}
\usage{
correctBsSeqCoverage(tumor, normal, bsgc = NULL, maps = NULL)
}
\arguments{
\item{tumor}{Binned tumor coverage - a GRanges from binCoverage}

\item{normal}{Binned normal coverage - a GRanges from binCoverage}

\item{bsgc}{Binned c2t GC content (DEFAULT: bsgc.hg19, 1 kb bins)}

\item{maps}{Binned c2t mappability (DEFAULT: maps.hg19, 1 kb bins)}
}
\value{
\preformatted{   Corrected tumor and normal read counts
}
}
\description{
Limits coverage to chromosomes 1 through 22 and X (for better or for worse)
}
\examples{

  tum_bed <- system.file("extdata", "MCF7_Cunha_chr11p15_shuffled.bed.gz",
                         package="biscuiteer")
  nor_bed <- system.file("extdata", "MCF7_Cunha_chr11p15.bed.gz",
                         package="biscuiteer")
  tum_vcf <- system.file("extdata", "MCF7_Cunha_shuffled_header_only.vcf.gz",
                         package="biscuiteer")
  nor_vcf <- system.file("extdata", "MCF7_Cunha_header_only.vcf.gz",
                         package="biscuiteer")

  tumor <- read.biscuit(BEDfile=tum_bed, VCFfile=tum_vcf, merged=FALSE)
  normal <- read.biscuit(BEDfile=nor_bed, VCFfile=nor_vcf, merged=FALSE)

  corr <- correctBsSeqCoverage(tumor=tumor, normal=normal)

}
\seealso{
binCoverage

TitanCNA::correctReadDepth
}
