% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctBsSeqCoverage.R
\name{correctBsSeqCoverage}
\alias{correctBsSeqCoverage}
\title{tweaked version of TitanCNA's preprocessing function; limits to chr{1:22,X}}
\usage{
correctBsSeqCoverage(tumor, normal, bsgc = NULL, maps = NULL)
}
\arguments{
\item{tumor}{binned tumor coverage (a GRanges from binCoverage)}

\item{normal}{binned normal coverage (a GRanges from binCoverage)}

\item{bsgc}{binned c2t GC content (default is bsgc.hg19, 1kb bins)}

\item{maps}{binned c2t mappability (default is maps.hg19, 1kb bins)}
}
\value{
\preformatted{     corrected tumor and normal read counts
}
}
\description{
tweaked version of TitanCNA's preprocessing function; limits to chr{1:22,X}
}
