% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctDepth.R
\name{correctDepth}
\alias{correctDepth}
\title{tweaked version of HMMcopy's GC bias and mappability function}
\usage{
correctDepth(x, mappability = 0.9, samplesize = 50000,
  verbose = TRUE)
}
\arguments{
\item{x}{binned hg19 coverage with GC content and mappability cols}

\item{mappability}{minimum mappability to use bins as examplars for others}

\item{samplesize}{lowess sample size (points to sample; default 50000)}

\item{verbose}{be chatty? (TRUE)}
}
\value{
\preformatted{       corrected tumor and normal read counts
}
}
\description{
tweaked version of HMMcopy's GC bias and mappability function
}
