% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGindex.R
\name{CpGindex}
\alias{CpGindex}
\title{simple index of hypermethylation-at-CpG-islands vs hypomethylation-in-WCGWs}
\usage{
CpGindex(x, CGIs = NULL, WCGW = NULL, PMDs = NULL)
}
\arguments{
\item{x}{a BSseq object}

\item{CGIs}{a GRanges of hypermethylation targets, or NULL (default CGIs)}

\item{WCGW}{a GRanges of hypomethylation targets, or NULL (default WCGW)}
}
\value{
\preformatted{   a list: `hyper`, `hypo`, and `ratio` (data.frame) by sample
}
}
\description{
The default for this function is to use the HMM-defined CpG islands from
Hao Wu's paper (Wu, Caffo, Jaffee, Irizarry & Feinberg, Biostatistics 2010)
as generic "hypermethylation" targets (obviously tissue-specific subsets
might be more informative on a number of levels, e.g. as bivalent sites),
and the solo-WCGW sites within common partially methylated domains from
Wanding Zhou and Huy Dinh's paper (Zhou, Dinh, et al, Nat Genetics 2018)
as genetic "hypomethylation" targets (as above, obvious caveats about tissue
specificity and user-supplied possibilities exist).
}
\details{
The return value of the function is an idiotically simple measure, comprised
of hyperCGI/hypoPMD (both the components and the ratio). The PMD "score" is
a base-coverage-weighted average of losses to solo-WCGW bases within PMDs;
the CGI score is similarly base-coverage-weighted but across HMM CGI CpGs.

By providing different targets a user can customize as needed.
}
