% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CpGindex.R
\name{CpGindex}
\alias{CpGindex}
\title{simple index of hypermethylation-at-CpG-islands vs hypomethylation-in-WCGWs}
\usage{
CpGindex(x, CpGi = NULL, WCGWs = NULL)
}
\arguments{
\item{x}{a BSseq object}

\item{CpGi}{a GRanges of hypermethylation targets, or NULL (default CpGis)}

\item{WCGWs}{a GRanges of hypomethylation targets, or NULL (deafult WCGWs)}
}
\value{
\preformatted{   a list: `hyper`, `hypo`, and `ratio` (data.frame) by sample
}
}
\description{
The default for this function is to use the HMM-defined CpG islands from
Hao Wu's paper (Wu, Caffo, Jaffee, Irizarry & Feinberg, Biostatistics 2010)
as generic "hypermethylation" targets (obviously tissue-specific subsets
might be more informative on a number of levels, e.g. as bivalent sites),
and the solo-WCGW sites within common partially methylated domains from
Wanding Zhou and Huy Dinh's paper (Zhou, Dinh, et al, Nat Genetics 2018)
as genetic "hypomethylation" targets (as above, obvious caveats about tissue
specificity and user-supplied possibilities exist).
}
\details{
The return value of the function is an idiotically simple measure, comprised
of hyper5mC/hypo5mC (both the components and the ratio). That's all: no more
and no less. By providing different targets a user can customize as needed.
}
