% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSR_subset.hg19}
\alias{ENSR_subset.hg19}
\title{ENSR_subset data from hg19 genome}
\usage{
data(ENSR_subset.hg19, package="biscuiteer")
}
\description{
Subset of ENSEMBL regulatory build regions for hg19 genome
}
\keyword{data}
