% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremality.R
\name{extremality}
\alias{extremality}
\title{compute the fraction of a Bernoulli variance achieved by a proportion}
\usage{
extremality(x, raw = FALSE)
}
\arguments{
\item{x}{an rectangular object with proportions in it}

\item{raw}{skip the correction pass? (FALSE)}
}
\value{
\preformatted{the extremality of each row (if more than one) of the object 
}
}
\description{
Note: this works (efficiently) on matrices and DelayedMatrix objects.
Also note: since it is posssible for "raw" extremality to be > 1,
the function does a second pass to correct for this.
}
\details{
extremal <- c(1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0)
milder <- c(0.8, 0.2, 0.1, 0.3, 0.7, 0.1, 0.4, 0.1, 0.1, 0.2)

extremality(extremal, raw=TRUE)
extremality(extremal, raw=FALSE)

FIXME: this breaks when handed a single-row DelayedArray.
}
