% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fexpit.R
\name{fexpit}
\alias{fexpit}
\title{Calculate expanded expit function}
\usage{
fexpit(x, sqz = 1e-06)
}
\arguments{
\item{x}{A vector of values between -Inf and +Inf}

\item{sqz}{The amount by which to 'squeeze' (DEFAULT: 0.000001)}
}
\value{
\preformatted{A vector of values between 0 and 1 inclusive
}
}
\description{
Helper function for calculating expanded expit
}
\examples{

  num <- rnorm(100, mean = 0, sd = 100)
  exp <- fexpit(num)

}
