% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLoci.R
\name{filterLoci}
\alias{filterLoci}
\title{Filter loci with zero coverage}
\usage{
filterLoci(bsseq, testCovariate)
}
\arguments{
\item{bsseq}{A bsseq object for filtering}

\item{testCovariate}{The name of the pData column dmrseq will test on}
}
\value{
\preformatted{          A bsseq object ready for dmrseq to use
}
}
\description{
Function potentially used to be a part of dmrseq. Included here to avoid
dmrseq failing due to any number of reasons related to lack of coverage.
}
\details{
The code is adapted from the precheck loop of dmrseq::dmrseq
}
\examples{

  tcga_mrg <- system.file("extdata",
                          "TCGA_BLCA_A13J_chr11p15_unmerged.bed.gz",
                          package = "biscuiteer")
  tcga_shf <- system.file("extdata",
                          "TCGA_BLCA_A13J_chr11p15_shuffled_unmerged.bed.gz",
                           package = "biscuiteer")
  tcga_mvcf <- system.file("extdata",
                           "TCGA_BLCA_A13J_header_only.vcf.gz",
                           package = "biscuiteer")
  tcga_svcf <- system.file("extdata",
                           "TCGA_BLCA_A13J_shuffled_header_only.vcf.gz",
                           package = "biscuiteer")
  bisc1 <- read.biscuit(BEDfile = tcga_mrg, VCFfile = tcga_mvcf,
                        merged = FALSE, genome = "hg38")
  bisc2 <- read.biscuit(BEDfile = tcga_shf, VCFfile = tcga_svcf,
                        merged = FALSE, genome = "hg38")

  comb <- unionize(bisc1, bisc2)

  filt <- filterLoci(comb, "sampleNames")

}
\seealso{
dmrseq

WGBSeq

RRBSeq
}
