% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLoci.R
\name{filterLoci}
\alias{filterLoci}
\title{filter out loci that have zero coverage for one or more conditions}
\usage{
filterLoci(bs, testCovariate, ...)
}
\arguments{
\item{bs}{a bsseq object for filtration}

\item{testCovariate}{the name of the pData column dmrseq will test on}

\item{...}{not currently used; may pass to dmrseq in the future}
}
\value{
\preformatted{           a bsseq object ready for dmrseq to use
}
}
\description{
This function used to be part of dmrseq, or so we recall. In any event,
it is a colossal drag to set up a dmrseq run only to discover that it fails
for seemingly mysterious reasons owing to lack of coverage.
}
\details{
The code is adapted from the precheck loop of dmrseq::dmrseq().
}
\seealso{
dmrseq

WGBSeq

RRBSeq
}
