% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNAs.R
\name{fixNAs}
\alias{fixNAs}
\title{helper function to play nice with BSseq's expectations about the M matrix}
\usage{
fixNAs(x, y = 0)
}
\arguments{
\item{x}{the matrix-like object containing NAs (holes) to fix}

\item{y}{the value with which to fill the holes (default is 0)}
}
\value{
x, but with holes filled, and as a matrix
}
\description{
helper function to play nice with BSseq's expectations about the M matrix
}
