% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNAs.R
\name{fixNAs}
\alias{fixNAs}
\title{Replace NAs with another value}
\usage{
fixNAs(x, y = 0, sparse = FALSE)
}
\arguments{
\item{x}{The matrix-like object containing NAs to fix}

\item{y}{The value to replace the NAs with (DEFAULT: 0)}

\item{sparse}{Make the result a Matrix object? (DEFAULT: FALSE)}
}
\value{
\preformatted{   x with no NAs (possibly a sparse Matrix)
}
}
\description{
Useful for coercing matrices into how bsseq is expecting the M matrix to be.
}
\examples{

  nom <- c(rep(c(1,4,NA,9,NA,NA,7,NA), 5))
  fixNAs(nom)

}
