% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flogit.R
\name{flogit}
\alias{flogit}
\title{Calculate squeezed logit function}
\usage{
flogit(x, sqz = 0.000001)
}
\arguments{
\item{x}{A vector of values between 0 and 1 inclusive}

\item{sqz}{The amount by which to 'squeeze' (DEFAULT: 0.000001)}
}
\value{
\preformatted{A vector of values between -Inf and +Inf
}
}
\description{
Helper function for calculating squeezed logit
}
