% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMvals.R
\name{getMvals}
\alias{getMvals}
\title{helper function for compartment inference (shrink-by-smoothing)}
\usage{
getMvals(bsseq, chrom = "chr22", minCov = 3, resolution = 100000,
  k = 1)
}
\arguments{
\item{bsseq}{a BSseq object with methylated and total reads}

\item{chrom}{what chromosome to grab ("chr22")}

\item{minCov}{minimum read coverage to use}

\item{resolution}{resolution for binning (1e5)}

\item{k}{pseudoreads for smoothing}
}
\value{
\preformatted{     a matrix of Mvalues
}
}
\description{
since we want something resembling M-values for compartment inference,
this function grabs suitably deep (>= 3 reads in >=1 sample) measurements
and turns them into lightly moderated M-values for compartment calling
by Dirichlet smoothing (adding addReads reads to each of M and Covg)
}
