% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogitFracMeth.R
\name{getMvals}
\alias{getMvals}
\title{Alias for getLogitFracMeth}
\usage{
getMvals(x, minCov = 3, minSamp = 2, k = 0.1, r = NULL)
}
\description{
Alias for getLogitFracMeth
}
\seealso{
getLogitFracMeth
}
