% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H9state23unmeth.hg19}
\alias{H9state23unmeth.hg19}
\title{H9state23unmeth.hg19}
\usage{
data(H9state23unmeth.hg19, package="biscuiteer")
}
\description{
Hypermethylated targets in bivalent histone sites from H9 embryonic stem
cells which were unmethylated across normal cells for hg19 genome
}
\details{
GRanges was generated by taking the HMM-derived CpG islands (described in
?HMM_CpG_islands.hg19) and overlapping with regions that were unmethylated in
normal H9 stem cells and had a ChromHMM state of 2 or 3 (see
https://www.nature.com/articles/nmeth.1906#MOESM194 for a description of
ChromHMM)
Return type: GRanges
}
\keyword{data}
