% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{H9state23unmeth.hg38}
\alias{H9state23unmeth.hg38}
\title{H9state23unmeth.hg38}
\usage{
data(H9state23unmeth.hg38, package="biscuiteer")
}
\description{
Hypermethylated targets in bivalent histone sites from H9 embryonic stem
cells which were unmethylated across normal cells for hg38 genome
}
\keyword{data}
