% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.biscuit.R
\name{load.biscuit}
\alias{load.biscuit}
\title{Alias for read.biscuit}
\usage{
load.biscuit(BEDfile, VCFfile, merged, sampleNames = NULL,
  simplify = FALSE, genome = "hg19", how = c("data.table", "readr"),
  hdf5 = FALSE, hdf5dir = NULL, sparse = FALSE, chunkSize = 1e+06,
  chr = NULL, which = NULL, verbose = FALSE)
}
\description{
Alias for read.biscuit
}
\seealso{
read.biscuit
}
