% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.biscuit.unmerged.R
\name{load.biscuit.unmerged}
\alias{load.biscuit.unmerged}
\title{a bsseq loader for Biscuit cg/ch output (BED-like format, 2 cols/sample)
e.g. P01-028-T06.joint.ch.hg19.bed.gz has 3 samples, and thus 9 columns}
\usage{
load.biscuit.unmerged(params)
}
\arguments{
\item{params}{parameters from having run checkBiscuitBED}
}
\value{
\preformatted{       a BSseq object from the bsseq package
}
}
\description{
a bsseq loader for Biscuit cg/ch output (BED-like format, 2 cols/sample)
e.g. P01-028-T06.joint.ch.hg19.bed.gz has 3 samples, and thus 9 columns
}
\seealso{
load.biscuit

checkBiscuitBED
}
\concept{
bsseq

BSseq
}
