% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHDF5.R
\name{loadHDF5}
\alias{loadHDF5}
\title{a simple local wrapper around HDF5Array::loadHDF5SummarizedExperiment}
\usage{
loadHDF5(dir = "HDF5")
}
\arguments{
\item{dir}{path to the file assays.h5 (and se.rds); default is "HDF5"}
}
\value{
\preformatted{   a SummarizedExperiment-derived object (such as a BSseq object)
}
}
\description{
a simple local wrapper around HDF5Array::loadHDF5SummarizedExperiment
}
