% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadHDF5.R
\name{loadHDF5}
\alias{loadHDF5}
\title{Load HDF5-backed SummarizedExperiment}
\usage{
loadHDF5(dir = "HDF5")
}
\arguments{
\item{dir}{Path to the assays.h5 (and se.rds) files (DEFAULT: "HDF5")}
}
\value{
\preformatted{A SummarizedExperiment-derived object (list a bsseq object)
}
}
\description{
Local wrapper around HDF5Array::loadHDF5SummarizedExperiment. NOTE: CURRENTLY
BISCUITEER DOES NOT HAVE HDF5-BACKING FUNCTIONALITY. USE AT YOUR OWN RISK!!!!
}
\seealso{
HDF5Array::loadHDF5SummarizedExperiment
}
