% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{make an in-core BSseq object from the results of reading a Biscuit BED file}
\usage{
makeBSseq(tbl, params)
}
\arguments{
\item{tbl}{a tibble (from read_tsv)}

\item{params}{parameters (from checkBiscuitBED)}
}
\value{
an in-core BSseq object
}
\description{
make an in-core BSseq object from the results of reading a Biscuit BED file
}
\seealso{
makeBSseq_HDF5
}
