% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{Make an in-memory bsseq object from a biscuit BED}
\usage{
makeBSseq(tbl, params, simplify = FALSE, verbose = FALSE)
}
\arguments{
\item{tbl}{A tibble (from read_tsv) or a data.table (from fread)}

\item{params}{Parameters from checkBiscuitBED}

\item{simplify}{Simplify sample names by dropping .foo.bar.hg19? (or
similar) (DEFAULT: FALSE)}

\item{verbose}{Print extra statements? (DEFAULT: FALSE)}
}
\value{
\preformatted{     An in-memory bsseq object
}
}
\description{
Beware that any reasonably large BED files may not fit into memory!
}
\examples{

  library(data.table)

  tcga_bed <- system.file("extdata", "TCGA_BLCA_A13J_chr11p15_merged.bed.gz",
                          package = "biscuiteer")
  tcga_vcf <- system.file("extdata", "TCGA_BLCA_A13J_header_only.vcf.gz",
                          package = "biscuiteer")

  params <- checkBiscuitBED(BEDfile = tcga_bed, VCFfile = tcga_vcf,
                            merged = TRUE, how = "data.table")

  select <- grep("\\\\.context", params$colNames, invert=TRUE)
  cmd <- paste("gunzip -c", params$tbx$path) # for mac compatibility
  tbl <- fread(cmd=cmd, sep="\\t", sep2=",", fill=TRUE, na.string=".", 
               select=select)
  if (params$hasHeader == FALSE) names(tbl) <- params$colNames[select]
  names(tbl) <- sub("^#", "", names(tbl))
  
  tbl <- tbl[rowSums(is.na(tbl)) == 0, ]
  bsseq <- makeBSseq(tbl = tbl, params = params)

}
