% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{Make an in-memory bsseq object from a biscuit BED}
\usage{
makeBSseq(tbl, params, simplify = FALSE, verbose = FALSE)
}
\arguments{
\item{tbl}{A tibble (from read_tsv) or a data.table (from fread)}

\item{params}{Parameters from checkBiscuitBED}

\item{simplify}{Simplify sample names by dropping .foo.bar.hg19? (or
similar) (DEFAULT: FALSE)}

\item{verbose}{Print extra statements? (DEFAULT: FALSE)}
}
\value{
\preformatted{     An in-memory bsseq object
}
}
\description{
Beware that any reasonably large BED files may not fit into memory!
}
\examples{

}
