% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{make an in-core BSseq object from a biscuit BED}
\usage{
makeBSseq(tbl, params, simplify = FALSE, verbose = FALSE)
}
\arguments{
\item{tbl}{a tibble (from read_tsv) or a data.table (from fread())}

\item{params}{parameters (from checkBiscuitBED)}

\item{simplify}{simplify sample names by dropping .foo.bar.hg19 & similar}

\item{verbose}{be verbose about what is happening? (FALSE)}
}
\value{
an in-core BSseq object
}
\description{
make an in-core BSseq object from a biscuit BED
}
\seealso{
makeBSseq_HDF5
}
