% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq.R
\name{makeBSseq}
\alias{makeBSseq}
\title{make an in-core BSseq object from the results of reading a Biscuit BED file}
\usage{
makeBSseq(dt, betacols, covgcols, pData, sparse = NULL)
}
\arguments{
\item{dt}{a data.table (from the BED file)}

\item{betacols}{the beta column names (from checkBiscuitBED)}

\item{covgcols}{the coverage column names (from checkBiscuitBED)}

\item{pData}{a DataFrame (usually from checkBiscuitBED)}

\item{sparse}{make the object Matrix-backed? (NULL; if beneficial)}
}
\value{
an in-core BSseq object
}
\description{
make an in-core BSseq object from the results of reading a Biscuit BED file
}
\seealso{
makeBSseq_HDF5
}
