% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBSseq_hdf5.R
\name{makeBSseq_hdf5}
\alias{makeBSseq_hdf5}
\title{make an HDF5-backed BSseq object from an (imported) Biscuit BED file}
\usage{
makeBSseq_hdf5(dt, betacols, covgcols, pData, sparse = NULL)
}
\arguments{
\item{dt}{a data.table (from the BED file)}

\item{betacols}{the beta column names}

\item{covgcols}{the coverage column names}

\item{pData}{a DataFrame (usually from checkBiscuitBED)}

\item{sparse}{make the object Matrix-backed? (NULL; if beneficial)}
}
\value{
an HDF5-backed BSseq object
}
\description{
make an HDF5-backed BSseq object from an (imported) Biscuit BED file
}
\seealso{
makeBSseq
}
