% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegs.R
\name{plotSegs}
\alias{plotSegs}
\title{Plot WGBScopy results}
\usage{
plotSegs(cn, segs = NULL, chr = "chr22", name = "Segmented", ...)
}
\arguments{
\item{cn}{Binned, corrected read depths from correctBsSeqCoverage}

\item{segs}{Segmentation result from WGBScopy}

\item{chr}{Chromosome to display (DEFAULT: "chr22")}

\item{name}{Name for the plot (DEFAULT: "Segmented")}

\item{...}{Other arguments to pass on to plot}
}
\value{
\preformatted{ Plot of WGBS copy results
}
}
\description{
Modified from HMMcopy
}
\examples{

  library(Homo.sapiens)

  tum_bed <- system.file("extdata", "MCF7_Cunha_chr11p15_shuffled.bed.gz",
                         package="biscuiteer")
  nor_bed <- system.file("extdata", "MCF7_Cunha_chr11p15.bed.gz",
                         package="biscuiteer")
  tum_vcf <- system.file("extdata", "MCF7_Cunha_shuffled_header_only.vcf.gz",
                         package="biscuiteer")
  nor_vcf <- system.file("extdata", "MCF7_Cunha_header_only.vcf.gz",
                         package="biscuiteer")

  tumor <- read.biscuit(BEDfile=tum_bed, VCFfile=tum_vcf, merged=FALSE)
  normal <- read.biscuit(BEDfile=nor_bed, VCFfile=nor_vcf, merged=FALSE)

  corr <- correctBsSeqCoverage(tumor=tumor, normal=normal)
  segs <- WGBSseg(cn=corr)
  plotSegs(corr, segs, "chr11")

}
