% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEpibed.R
\name{readEpibed}
\alias{readEpibed}
\title{Read in and decode the RLE representation of
the epibed format out of biscuit epiallele
returning a read-level GRanges object.}
\usage{
readEpibed(epibed, is.nome = FALSE, genome = NULL)
}
\arguments{
\item{epibed}{The path to the epibed file}

\item{is.nome}{Whether the epibed format is derived from NOMe-seq or not}

\item{genome}{What genome did this come from (e.g. 'hg19')}
}
\value{
A GRanges object
}
\description{
NOTE: it is recommended that you ran biscuit epiallele
with the -g option enabled to subset to a specific region.
If not, this will be the equivalent of reading in the entire
BAM. Future work will be to subset to definied regions.
}
\examples{
epibed.nome <- system.file("extdata", "hct116.nome.epiread",
                           package="biscuiteer")
epibed.bsseq <- system.file("extdata", "hct116.bsseq.epiread",
                            package="biscuiteer")
epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
                             genome = "hg19")
epibed.bsseq.gr <- readEpibed(epibed = epibed.bsseq,
                              genome = "hg19")
}
