% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRBSeq.R
\name{RRBSeq}
\alias{RRBSeq}
\title{wrapper for \link{e}RRBS settings appropriate to dmrseq}
\usage{
RRBSeq(bs, testCovariate, cutoff = 0.2, bpSpan = 750, ...)
}
\arguments{
\item{bs}{a bsseq object}

\item{testCovariate}{the pData column to test on}

\item{cutoff}{the minimum CpG-wise difference to use (0.2)}

\item{bpSpan}{span of smoother AND max gap in DMR CpGs (750)}

\item{...}{other arguments to pass along to dmrseq}
}
\value{
\preformatted{           a GRanges object (same as from dmrseq)
}
}
\description{
wrapper for \link{e}RRBS settings appropriate to dmrseq
}
