% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveHDF5.R
\name{saveHDF5}
\alias{saveHDF5}
\title{Save HDF5-backed SummarizedExperiment}
\usage{
saveHDF5(x, dir = "HDF5", ...)
}
\arguments{
\item{x}{The SummarizedExperiment-derived object to save as HDF5}

\item{dir}{Path to store the HDF5 file "assays.h5" (DEFAULT: "HDF5")}

\item{...}{Arguments to pass to HDF5Array::saveHDF5SummarizedExperiment}
}
\value{
\preformatted{Invisible. The HDF5-backed version of the object being saved
}
}
\description{
Local wrapper around HDF5Array::saveHDF5SummarizedExperiment: NOTE: CURRENTLY
BISCUITEER DOES NOT HAVE HDF5-BACKING FUNCTIONALITY. USE AT YOUR OWN RISE!!!!
}
