% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifySampleNames.R
\name{simplifySampleNames}
\alias{simplifySampleNames}
\title{Simplify bsseq sample names}
\usage{
simplifySampleNames(x)
}
\arguments{
\item{x}{A SummarizedExperiment-derived object, or a character vector}
}
\value{
The input object, but with simplified sample names
}
\description{
Tries using the longest common subsequence to figure out what can be
dropped. Usually used for VCF columns.
}
\examples{

  tcga_bed <- system.file("extdata", "TCGA_BLCA_A13J_chr11p15_merged.bed.gz",
                          package = "biscuiteer")
  tcga_vcf <- system.file("extdata", "TCGA_BLCA_A13J_header_only.vcf.gz",
                          package = "biscuiteer")
  bisc <- read.biscuit(BEDfile = tcga_bed, VCFfile = tcga_vcf,
                       merged = TRUE, genome = "hg38")

  bisc <- simplifySampleNames(bisc)

}
