% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBsSeqOver.R
\name{summarizeBsSeqOver}
\alias{summarizeBsSeqOver}
\title{Pretty much what it says on the tin.  Summarize methylation over regions.}
\usage{
summarizeBsSeqOver(bsseq, segs, dropNA = TRUE, impute = FALSE)
}
\arguments{
\item{bsseq}{the BSseq object to summarize}

\item{segs}{the regions to summarize over}

\item{dropNA}{whether to drop rows with > half of samples NaN (TRUE)}

\item{impute}{whether to impute NAs (FALSE)}
}
\value{
matrix    a matrix of regional methylation fractions
}
\description{
Pretty much what it says on the tin.  Summarize methylation over regions.
}
