% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBsSeqOver.R
\name{summarizeBsSeqOver}
\alias{summarizeBsSeqOver}
\title{like it says on the tin...}
\usage{
summarizeBsSeqOver(bsseq, segs, dropNA = TRUE, impute = FALSE)
}
\arguments{
\item{bsseq}{the BSseq object to summarize}

\item{segs}{the regions to summarize over}

\item{dropNA}{whether to drop rows where > 50% of samples are NaN (TRUE)}

\item{impute}{whether to impute NAs (FALSE)}
}
\value{
matrix    a matrix of regional "beta values"
}
\description{
like it says on the tin...
}
