% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeBsSeqOver.R
\name{summarizeBsSeqOver}
\alias{summarizeBsSeqOver}
\title{Pretty much what it says on the tin.}
\usage{
summarizeBsSeqOver(bsseq, segs, dropNA = FALSE, impute = FALSE)
}
\arguments{
\item{bsseq}{the BSseq object to summarize}

\item{segs}{regions to summarize over (a GRanges; no GRangesLists yet)}

\item{dropNA}{whether to drop rows with > half of samples NaN (FALSE)}

\item{impute}{whether to impute NAs/NaNs (FALSE)}
}
\value{
matrix    a matrix of regional methylation fractions
}
\description{
Summarize methylation over regions. Mostly a wrapper for getMeth.
}
