% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabixToGr.R
\name{tabixRetrieve}
\alias{tabixRetrieve}
\title{Read from tabix-indexed bed file to list objects}
\usage{
tabixRetrieve(
  paths,
  chr,
  start = 1,
  end = 2^28,
  sample_names = NULL,
  is.epibed = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{paths}{path(s) to the bed files}

\item{chr}{chromosome name}

\item{start}{start coordinate of region of interest}

\item{end}{end coordinate of region of interest}

\item{sample_names}{sample names, just use paths if not specified}

\item{is.epibed}{whether the input is epibed format}

\item{BPPARAM}{how to parallelize}
}
\value{
a list object with DNA methylation level and depth
}
\description{
Read from tabix-indexed bed file to list objects
}
