% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epistate.R
\name{tabulateEpibed}
\alias{tabulateEpibed}
\title{Tabulate to a read or fragment level matrix for
CGH or GCH context methylation. The input for this
function is the GRanges output from readEpibed().
It will auto-detect whether this a BS-seq or a
NOMe-seq run and tabulate accordingly.}
\usage{
tabulateEpibed(gr)
}
\arguments{
\item{gr}{The epibed GRanges object from readEpibed()}
}
\value{
A matrix or list of matrices
}
\description{
Tabulate to a read or fragment level matrix for
CGH or GCH context methylation. The input for this
function is the GRanges output from readEpibed().
It will auto-detect whether this a BS-seq or a
NOMe-seq run and tabulate accordingly.
}
\examples{
epibed.nome <- system.file("extdata", "hct116.nome.epiread",
                           package="biscuiteer")
epibed.nome.gr <- readEpibed(epibed = epibed.nome, is.nome = TRUE,
                             genome = "hg19")
epibed.tab.nome <- tabulateEpibed(epibed.nome.gr)
}
