% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_WGBS_CNA.R
\name{TN_WGBS_CNA}
\alias{TN_WGBS_CNA}
\title{Wrapper for matched tumor-normal bigWig coverage processing & segmentation}
\usage{
TN_WGBS_CNA(stub0, stub1 = NULL, bsgc = NULL, maps = NULL, ...)
}
\arguments{
\item{stub0}{bigWig stub for normal: paste0(stub0, ".covg.hg19.bw")}

\item{stub1}{bigWig stub for tumor: paste0(stub1, ".covg.hg19.bw") \link{NULL}}

\item{bsgc}{binned c2t GC content (default is bsgc.hg19, 1k bins) \link{NULL}}

\item{maps}{binned c2t mappability (default is maps.hg19, 1k bins)\link{NULL}}

\item{...}{other parameters passed to WGBSseg}
}
\value{
\preformatted{     a GRanges
}
}
\description{
Wrapper for matched tumor-normal bigWig coverage processing & segmentation
}
