% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN_WGBS_CNA.R
\name{TN_WGBS_CNA}
\alias{TN_WGBS_CNA}
\title{Matched tumor-normal coverage processing & segmentation}
\usage{
TN_WGBS_CNA(stub0, stub1 = NULL, bsgc = NULL, maps = NULL, ...)
}
\arguments{
\item{stub0}{bigWig stub for normal: paste0(stub0, ".covg.hg19.bw")}

\item{stub1}{bigWig stub for tumor: paste0(stub1, ".covg.hg19.bw") \link{NULL}}

\item{bsgc}{binned c2t GC content (default is bsgc.hg19, 1k bins) \link{NULL}}

\item{maps}{binned c2t mappability (default is maps.hg19, 1k bins)\link{NULL}}

\item{...}{other parameters passed to WGBSseg}
}
\value{
\preformatted{     a GRanges
}
}
\description{
Matched tumor-normal coverage processing & segmentation
}
\examples{
  stubs <- c("P01-010", "P01-012", "P01-015", "P01-016", "P01-017")
  stubs <- c(stubs, "P01-018", "P01-020", "P01-021")
  names(stubs) <- paste0(stubs, "-T06-T-P01")
  names(stubs)[1] <- "P01-010-T07-T-P01"
  segs <- list()
  for(i in names(stubs)) segs[[i]] <- TN_WGBS_CNA(stubs[[i]])$segs # robust
  segs <- GRangesList(segs) # to feed to grToSeg and get full .seg output
  for(i in names(segs)) export(segs[[i]], paste0(i, ".CN.hg19.bw"))

}
