% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionize.R
\name{unionize}
\alias{unionize}
\title{Combine bsseq objects together without losing information}
\usage{
unionize(bs1, ..., parallel = FALSE, onlyChrs = NULL)
}
\arguments{
\item{bs1}{A bsseq object (returned unaltered if length(list(...)) = 0)}

\item{...}{One or more bsseq objects to combine with bs1}

\item{parallel}{Split the bsseq objects by chrom and parallelize
(DEFAULT: FALSE)}

\item{onlyChrs}{Retain a specific subset of chromosomes - if NULL, keep all
(DEFAULT: NULL)}
}
\value{
\preformatted{     A larger and more sparse bsseq object
}
}
\description{
Takes provided bsseq objects, the union of their GRanges, fills out the
sites no in the union with 0M/0Cov (or ./. in a sparse Matrix), and returns
the even-sparser bsseq holding all of them.
}
\details{
All conditions governing cbind(bsseq1, bsseq2, ...) also apply to unionize.
}
\examples{

}
\seealso{
unionizeBSseq
}
