% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionize.R
\name{unionize}
\alias{unionize}
\title{take a bunch of bsseq objects, take the union of their granges(),
pad out un-covered sites with 0M/0Cov (i.e., ./. in a sparse Matrix),
and return the now-even-sparser bsseq holding all of them. All
conditions governing cbind(bsseq1, bsseq2) also apply to unionize().}
\usage{
unionize(bs1, ..., parallel = TRUE)
}
\arguments{
\item{bs1}{a bsseq object (return unaltered if length(list(...)) == 0)}

\item{...}{one or more bsseq objects to combine with the first one}

\item{parallel}{split the bsseq objects by chrom and parallelize? (TRUE)}
}
\value{
\preformatted{   a larger and more sparse bsseq object
}
}
\description{
take a bunch of bsseq objects, take the union of their granges(),
pad out un-covered sites with 0M/0Cov (i.e., ./. in a sparse Matrix),
and return the now-even-sparser bsseq holding all of them. All
conditions governing cbind(bsseq1, bsseq2) also apply to unionize().
}
