% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionizeBSseq.R
\name{unionizeBSseq}
\alias{unionizeBSseq}
\title{Combine two bsseq objects together}
\usage{
unionizeBSseq(bs1, bs2, biggrl, what = c("M", "Cov"))
}
\arguments{
\item{bs1}{bsseq object number 1}

\item{bs2}{bsseq object number 2}

\item{biggrl}{A GRangesList with (bs1-only, both, and bs2-only) regions}

\item{what}{"M" (methylation) or "Cov" (coverage)}
}
\value{
\preformatted{   A Matrix
}
}
\description{
Takes two bsseq objects and the overlaps of the union of their rowRanges,
fills the sites not in the union with 0M/0Cov (or ./. in a sparse Matrix),
then returns a Matrix containing the joint information from both objects.
}
\examples{

}
\seealso{
unionize
}
