% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unionizeBSseq.R
\name{unionizeBSseq}
\alias{unionizeBSseq}
\title{Smoosh a couple of BSseq objects together, sensibly.}
\usage{
unionizeBSseq(bs1, bs2, biggrl, what = c("M", "Cov"), parallel = TRUE)
}
\arguments{
\item{bs1}{a BSseq object}

\item{bs2}{another BSseq object}

\item{biggrl}{a GRangesList with (bs1-only, both, bs2-only) regions}

\item{what}{"M" (methylation) or "Cov" (coverage)}
}
\value{
a DelayedMatrix
}
\description{
take two bsseq objects and the overlaps of the union of their rowRanges,
pad out un-covered sites with 0M/0Cov (i.e., ./. in a sparse Matrix),
and return a Matrix containing the joint information from both.
}
