% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSage.R
\name{WGBSage}
\alias{WGBSage}
\title{Guess ages using Horvath-style 'clock' models}
\usage{
WGBSage(bsseq, model = c("horvath", "horvathshrunk", "hannum",
  "skinandblood"), padding = 15, useENSR = FALSE, useHMMI = FALSE,
  minCovg = 5, impute = FALSE, minSamp = 5, genome = NULL,
  dropBad = FALSE, ...)
}
\arguments{
\item{bsseq}{A bsseq object (must have assays named \code{M} and \code{Cov})}

\item{model}{Which model ("horvath", "horvathshrunk", "hannum",
"skinandblood")}

\item{padding}{How many bases +/- to pad the target CpG by (DEFAULT: 15)}

\item{useENSR}{Use ENSEMBL regulatory region bounds instead of CpGs
(DEFAULT: FALSE)}

\item{useHMMI}{Use HMM CpG island boundaries instead of padded CpGs
(DEFAULT: FALSE)}

\item{minCovg}{Minimum regional read coverage desired to estimate 5mC
(DEFAULT: 5)}

\item{impute}{Use k-NN imputation to fill in low-coverage regions?
(DEFAULT: FALSE)}

\item{minSamp}{Minimum number of non-NA samples to perform imputation
(DEFAULT: 5)}

\item{genome}{Genome to use as reference, if no genome(bsseq) is set
(DEFAULT: NULL)}

\item{dropBad}{Drop rows/cols with > half missing pre-imputation?
(DEFAULT: FALSE)}

\item{...}{Arguments to be passed to impute.knn, such as rng.seed}
}
\value{
\preformatted{    A list: call, methylation estimates, coefs, age (estimates)
}
}
\description{
See Horvath, Genome Biology, 2013 for more information
}
\details{
Note: the accuracy of the prediction will increase or decrease depending on
how various hyper-parameters are set by the user. This is NOT a hands-off
procedure, and the defaults are only a starting point for exploration. It
will not be uncommon to tune \code{padding}, \code{minCovg}, and \code{minSamp} for each
WGBS or RRBS experiment (and the latter may be impacted by whether dupes are
removed prior to importing data). Consider yourself forewarned. In the near
future we may add support for arbitrary region-coefficient inputs and result
transformation functions, which of course will just make the problems worse.

Also, please cite the appropriate papers for the Epigenetic Clock(s) you use:

For the 'horvath' or 'horvathshrunk' clocks, cite Horvath,
Genome Biology 2013.
For the 'hannum' clock, cite Hannum et al, Molecular Cell 2013.
For the 'skinandblood' clock, cite Horvath et al, Aging 2018.

Last but not least, keep track of the parameters YOU used for YOUR estimates.
The \code{call} element in the returned list of results is for this exact purpose.
If you need to recover the GRanges object used to average (or impute) DNAme
values for the model, try \code{as.character(rownames(result$meth))} on a result.
The coefficients for each of these regions are stored in result$coefs, and
the age estimates are stored in result$age (named, in case dropBad == TRUE).
}
\examples{

  tcga_mrg <- system.file("extdata",
                          "TCGA_BLCA_A13J_chr11p15_unmerged.bed.gz",
                          package = "biscuiteer")
  tcga_shf <- system.file("extdata",
                          "TCGA_BLCA_A13J_chr11p15_shuffled_unmerged.bed.gz",
                           package = "biscuiteer")
  tcga_mvcf <- system.file("extdata",
                           "TCGA_BLCA_A13J_header_only.vcf.gz",
                           package = "biscuiteer")
  tcga_svcf <- system.file("extdata",
                           "TCGA_BLCA_A13J_shuffled_header_only.vcf.gz",
                           package = "biscuiteer")
  bisc1 <- read.biscuit(BEDfile = tcga_mrg, VCFfile = tcga_mvcf,
                        merged = FALSE, genome = "hg38")
  bisc2 <- read.biscuit(BEDfile = tcga_shf, VCFfile = tcga_svcf,
                        merged = FALSE, genome = "hg38")

  comb <- unionize(bisc1, bisc2)
  ages <- WGBSage(comb, "horvath")

}
