% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSeq.R
\name{WGBSeq}
\alias{WGBSeq}
\title{wrapper for WGBS settings appropriate to dmrseq}
\usage{
WGBSeq(bs, testCovariate, bpSpan = 1000, ...)
}
\arguments{
\item{bs}{a bsseq object}

\item{testCovariate}{the pData column to test on}

\item{bpSpan}{span of smoother AND 2x max gap in DMR CpGs (1000)}

\item{...}{other arguments to pass along to dmrseq}
}
\value{
\preformatted{           a GRanges object (same as from dmrseq)
}
}
\description{
wrapper for WGBS settings appropriate to dmrseq
}
