% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSseq.R
\name{WGBSseg}
\alias{WGBSseg}
\title{Circular binary segmentation for GC- and mappability-corrected WGBS data}
\usage{
WGBSseg(cn, minAbs = 0.3, minWidth = 20000, ...)
}
\arguments{
\item{cn}{binned, corrected read depths from correctCoverage()}

\item{minAbs}{minimum absolute value of a segment to keep it (0.1)}

\item{minWidth}{minimum width of segment (NOT A GAP) to keep (25000)}

\item{...}{other parameters passed to fastseg}
}
\value{
\preformatted{         a GRanges (hg19 seqinfo)
}
}
\description{
FIXME: use some form of smoothing to glob segments together and THEN filter
on length (as opposed to right now where it is done rather roughly)
FIXME: use findOverlaps(resize(segs, width(segs) + minWidth, fix="center"))
}
