% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSseg.R
\name{WGBSseg}
\alias{WGBSseg}
\title{Copy number segmentation of GC- & mapping-corrected WGBS coverage}
\usage{
WGBSseg(cn, minAbs = 0.3, minWidth = 20000, ...)
}
\arguments{
\item{cn}{binned, corrected read depths from correctCoverage()}

\item{minAbs}{minimum absolute value of a segment to keep it (0.1)}

\item{minWidth}{minimum width of segment (NOT A GAP) to keep (25000)}

\item{...}{other parameters passed to fastseg}
}
\value{
\preformatted{         a GRanges (hg19 seqinfo)
}
}
\description{
FIXME: use some form of smoothing to glob segments together and THEN filter
on length (as opposed to right now where it is done rather roughly)
FIXME: use findOverlaps(resize(segs, width(segs) + minWidth, fix="center"))
}
\examples{

  library(fastseg)
  library(HMMcopy) 
  library(rtracklayer)
  files <- list.files(patt="P01-022-.*covg.hg19.bw") 
  names(files) <- sapply(strsplit(files, "(\\\\.|\\\\-)"), `[`, 3)
  covgs <- sapply(files, import)

  RMS1 <- correctBsSeqCoverage(covgs$tumor1, covgs$normal)
  RMS1seg <- WGBSseg(RMS1)
  plotSegs(RMS1, RMS1seg, "chr10")

  RMS2 <- correctBsSeqCoverage(covgs$tumor2, covgs$normal)
  RMS2seg <- WGBSseg(RMS2) 
  plotSegs(RMS2, RMS2seg, "chr10")

}
