% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGBSseg.R
\name{WGBSseg}
\alias{WGBSseg}
\title{Copy number segmentation of GC- and mapping-corrected WGBS coverage}
\usage{
WGBSseg(cn, minAbs = 0.3, minWidth = 20000, ...)
}
\arguments{
\item{cn}{Binned, corrected read depths from correctBsSeqCoverage}

\item{minAbs}{Minimum absolute value of a segment to keep (DEFAULT: 0.3)}

\item{minWidth}{Minimum width of segment (NOT A GAP) to keep
(DEFAULT: 20 000)}

\item{...}{Other parameters passed to fastseg}
}
\value{
\preformatted{     A GRanges (hg19 seqinfo)
}
}
\description{
Copy number segmentation of GC- and mapping-corrected WGBS coverage
}
\examples{

  library(Homo.sapiens)

  tum_bed <- system.file("extdata", "MCF7_Cunha_chr11p15_shuffled.bed.gz",
                         package="biscuiteer")
  nor_bed <- system.file("extdata", "MCF7_Cunha_chr11p15.bed.gz",
                         package="biscuiteer")
  tum_vcf <- system.file("extdata", "MCF7_Cunha_shuffled_header_only.vcf.gz",
                         package="biscuiteer")
  nor_vcf <- system.file("extdata", "MCF7_Cunha_header_only.vcf.gz",
                         package="biscuiteer")

  tumor <- read.biscuit(BEDfile=tum_bed, VCFfile=tum_vcf, merged=FALSE)
  normal <- read.biscuit(BEDfile=nor_bed, VCFfile=nor_vcf, merged=FALSE)

  corr <- correctBsSeqCoverage(tumor=tumor, normal=normal)
  segs <- WGBSseg(cn=corr)

}
